requirements.txt
Pillow==10.4.0
PyYAML==6.0.1
anndata==0.10.8
click==8.1.7
kaleido==0.2.1
matplotlib==3.7.1
networkx==3.3
numba==0.57.0
numpy==1.23.5
opencv_python==4.10.0.84
optuna==3.6.1
pandas==2.2.2
parquet==1.3.1
plotly==5.22.0
pyarrow==16.1.0
scanpy==1.10.2
scikit_learn==1.3.0
scikit-learn-extra==0.3.0
scikit_plot==0.3.7
scipy==1.14.0
seaborn==0.13.2
tensorboard==2.17.0
torch==2.3.1
torcheval==0.0.7
torchvision==0.18.1
umap_learn==0.5.6
